GH 4937: Edits the deadlock prevention thead logging.
Moves the logging after the thead sleep, so it only logs if it will try to interrupt the owner thread.
It also logs the owner stack trace if there is one to help detecting deadlocks.
Removes second call of the deadlock prevention thread in made in the method forceCloseActiveOperations since there is one already started from the close method before that.